+++ packages/app/src/features/dialogs/collateral/components/CollateralOverviewPanel.tsx
@@ -1,5 +1,5 @@
 import { TokenWithBalance } from '@/domain/common/types'
-import { TransactionOverview } from '@/ui/organisms/new/transaction-overview/TransactionOverview'
+import { TransactionOverview } from '@/ui/organisms/transaction-overview/TransactionOverview'
 import BigNumber from 'bignumber.js'
 
 export interface CollateralOverviewPanelProps {
