+++ packages/app/src/features/farm-details/dialogs/unstake/components/transaction-overview/TransactionOverview.tsx
@@ -1,7 +1,7 @@
 import { Token } from '@/domain/types/Token'
 import { HorizontalScroll } from '@/ui/atoms/horizontal-scroll/HorizontalScroll'
-import { TransactionOverview } from '@/ui/organisms/new/transaction-overview/TransactionOverview'
-import { RouteItem } from '@/ui/organisms/new/transaction-overview/rows/TransactionOverviewRoute'
+import { TransactionOverview } from '@/ui/organisms/transaction-overview/TransactionOverview'
+import { RouteItem } from '@/ui/organisms/transaction-overview/rows/TransactionOverviewRoute'
 import { testIds } from '@/ui/utils/testIds'
 import { assert } from '@/utils/assert'
 import { TxOverview } from '../../logic/createTxOverview'
