+++ packages/app/src/ui/organisms/transaction-overview/rows/TransactionOverviewStakedInFarm.tsx
