+++ packages/app/src/ui/organisms/transaction-overview/rows/TransactionOverviewTokenAmount.tsx
